relationship-extraction/requirements.txt
transformers>=4.30.0
torch>=2.0.0
pyyaml>=6.0
